refactor: extract inventory summary display to output.inventory
