+++ src/cja_auto_sdr/output/inventory/__init__.py
@@ -1,0 +1,7 @@
+"""Inventory output renderers — console display and JSON summary writing."""
+
+from cja_auto_sdr.output.inventory.summary import display_inventory_summary
+
+__all__ = [
+    "display_inventory_summary",
+]
